pyproject.toml
[project]
name = "azsearch-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.6.0b12",
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.21",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.8",
    "openai>=1.84.0",
    "requests>=2.32.3",
]
